before security

+++ src/main/java/Projet/ete/Parapharmcie/model/OrderCustomer.java
@@ -17,7 +17,7 @@ public class OrderCustomer {
     private Long id;
     private Date orderDate;
     private String deliveryAddress;
-    private double totalAmount;
+    private Integer totalAmount;
     @Enumerated
     private OrderStatus orderStatus;
